requirements.txt
Django
django-celery
